# packages/contracts/openapi/platform-spine.yaml
openapi: 3.0.3
info:
  title: platform-spine
  version: "1.0"
  description: Identity/RBAC, Job Orchestration state machine, Notifications, Gateway routing
paths:
  /v1/auth/register:
    post:
      summary: Register a new user and receive a JWT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                phone: { type: string }
                role: { type: string, enum: [customer, contractor, resource_owner], default: customer, description: "admin cannot be self-registered; it is created out-of-band (services/platform-spine/app/bootstrap.py)" }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  role: { type: string, enum: [customer, contractor, admin, resource_owner] }
        '409':
          description: Email already registered
        '403':
          description: CONTRACTOR_REGISTRATION_CLOSED - public sign-up for the contractor role is disabled (ALLOW_PUBLIC_CONTRACTOR_REGISTRATION=0); an operator creates these accounts
        '422':
          description: Validation error (including a role that cannot be self-registered, e.g. admin, or a password over 72 bytes)
        '429':
          description: Too many registration attempts from this client; see Retry-After
  /v1/auth/login:
    post:
      summary: Authenticate and receive a JWT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  role: { type: string, enum: [customer, contractor, admin, resource_owner] }
        '401':
          description: Invalid credentials
        '429':
          description: Too many login attempts from this client; see Retry-After
  /v1/jobs:
    post:
      summary: Create a new job (lead). Customer role only.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [location, job_type]
              properties:
                location:
                  type: object
                  properties:
                    lat: { type: number }
                    lng: { type: number }
                job_type: { type: string, enum: [residential, agricultural, commercial] }
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Only customers may create jobs
    get:
      summary: List jobs visible to the current user (own jobs for customers, all jobs for contractor/admin)
      parameters:
        - name: limit
          in: query
          required: false
          description: Max rows to return. Omit for the full list (unchanged default behavior).
          schema: { type: integer, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          required: false
          description: Rows to skip. Ignored unless limit is also set.
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
  /v1/jobs/{job_id}:
    get:
      summary: Get a single job's current state
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
  /v1/jobs/{job_id}/status:
    patch:
      summary: Advance a job to the next lifecycle status. Contractor/admin only, sequential transitions only.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [lead, site_location, requirement, estimation, price_calculation, quotation, customer_approval, booking, resource_allocation, drilling, progress, completion, payment, service_history]
      responses:
        '200':
          description: Job status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid transition (skips a state or moves backward)
        '403':
          description: Only contractor/admin may advance job status
  /v1/jobs/{job_id}/completion:
    post:
      summary: |
        Log actual depth and cost at job completion (FR-TRACK-03). Contractor
        only. Only valid when job status is "completion". Only writable once -
        a second POST returns 409 ALREADY_COMPLETED (SRS section 5 entity).
        Fetches the approved quotation total from the quotation service to
        compute variance at write time (BR-04). Emits job.completed event.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actual_depth_ft, actual_cost]
              properties:
                actual_depth_ft:
                  type: number
                  description: Actual depth drilled, in feet. If > quotation max_ft, depth overage charge is computed (BR-05).
                actual_cost:
                  type: number
                  description: Actual total cost incurred by the contractor.
      responses:
        '201':
          description: Completion record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCompletionRecord'
        '400':
          description: Job is not at completion status, or required fields missing
        '403':
          description: Contractor only
        '404':
          description: Job not found
        '409':
          description: Completion record already exists for this job (only writable once)

  /v1/jobs/{job_id}/completion/result:
    get:
      summary: Read the completion record and variance for a completed job (FR-TRACK-04).
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Completion record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCompletionRecord'
        '403':
          description: Can only view your own job's completion record
        '404':
          description: No completion record for this job yet

components:
  schemas:
    JobCompletionRecord:
      type: object
      description: |
        Written once at job completion (FR-TRACK-03). Stores actual depth/cost
        and the variance vs quoted total (BR-04). Never updated after creation.
      properties:
        job_id: { type: string, format: uuid }
        actual_depth_ft: { type: number }
        actual_cost: { type: string, format: decimal, description: "Decimal string, same serialization as quotation total_estimate" }
        quoted_total: { type: string, format: decimal, description: "Approved quotation total at time of completion" }
        variance: { type: string, format: decimal, description: "actual_cost minus quoted_total (negative = under budget)" }
        depth_overage_ft: { type: number, description: "Feet beyond quotation max_ft, or 0 if within range (BR-05)" }
        depth_overage_charge: { type: string, format: decimal, description: "Overage fee at contractor's configured rate, or 0 (BR-05)" }
        completed_at: { type: string, format: date-time }

    Job:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string }
        job_type: { type: string }
        location:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        created_at: { type: string, format: date-time }